pyproject.toml
[project]
name = "tanulmanyi-versenyek"
version = "0.1.0"
description = ""
authors = [
    {name = "Csaba Fülöp",email = "csaba.fulop@ericsson.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3,<7.0.0",
    "playwright>=1.57.0,<2.0.0",
    "pandas>=2.3.3,<3.0.0",
    "openpyxl>=3.1.5,<4.0.0"
]

[project.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "tanulmanyi_versenyek", from = "src"}]
